# server/src/main/resources/rest/openapi.yaml
openapi: 3.1.1
info:
  title: Evidence Engine REST Interface - OpenAPI 3.1.1
  description: |-
    This is a REST API for exporting lists of scientists, publications, etc., based on the OpenAPI 3.1.1 specification.
    Some useful links:
    - [The Evidence Engine page](https://campaign-resources.org/evidence-engine.html)
  termsOfService: https://campaign-resources.org/evidence-engine.html
  contact:
    email: admin@campaign-resources.org
  license:
    name: GNU Affero General Public License v3.0
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
  version: 0.0.1
externalDocs:
  description: Find out more about Evidence Engine
  url: https://campaign-resources.org/files/Evidence%20Engine/Evidence%20Engine.pdf
servers:
  - url: http://localhost:8080/rest
    description: Local development server
  - url: https://ee.campaign-resources.org/rest
    description: Production server
security: []
tags:
  - name: HTML
    description: Exported list in HTML format
    externalDocs:
      description: Find out more
      url: https://campaign-resources.org/evidence-engine.html#export-html
  - name: PDF
    description: Exported list in PDF format
    externalDocs:
      description: Find out more
      url: https://campaign-resources.org/evidence-engine.html#export-pdf
  - name: CSV
    description: Exported list in CSV format
    externalDocs:
      description: Find out more
      url: https://campaign-resources.org/evidence-engine.html#export-csv
  - name: RIS
    description: Exported list in RIS format. Applies only to recordKind = publications.
    externalDocs:
      description: Find out more
      url: https://campaign-resources.org/evidence-engine.html#export-ris
paths:
  /export/{recordKind}:
    get: 
      tags:
        - CSV
        - HTML
        - PDF
        - RIS
      summary: Exports a list
      description: Exports an optionally filtered list to CSV, HTML, PDF or RIS
      operationId: export
      parameters:
        - $ref: '#/components/parameters/recordKind'
        - $ref: '#/components/parameters/recordIdOpt'
        - $ref: '#/components/parameters/contentTypeOpt'
        - $ref: '#/components/parameters/textOpt'
        - $ref: '#/components/parameters/advancedSearchOpt'
        - $ref: '#/components/parameters/statusOpt'
        - $ref: '#/components/parameters/fromEntityIdOpt'
        - $ref: '#/components/parameters/fromEntityKindOpt'
        - $ref: '#/components/parameters/toEntityIdOpt'
        - $ref: '#/components/parameters/toEntityKindOpt'
        - $ref: '#/components/parameters/topicIdOpt'
        - $ref: '#/components/parameters/recursiveOpt'
        - $ref: '#/components/parameters/targetKindOpt'
        - $ref: '#/components/parameters/targetIdOpt'
        - $ref: '#/components/parameters/parentIdOpt'
        - $ref: '#/components/parameters/userIdOpt'
        - $ref: '#/components/parameters/fromOpt'
        - $ref: '#/components/parameters/toOpt'
        - $ref: '#/components/parameters/entityKindOpt'
        - $ref: '#/components/parameters/entityIdOpt'
        - $ref: '#/components/parameters/transactionKindOpt'
        - $ref: '#/components/parameters/pageNumberOpt'
        - $ref: '#/components/parameters/pageSizeOpt'
        - $ref: '#/components/parameters/sortOpt'
        - $ref: '#/components/parameters/columnsOpt'
        - $ref: '#/components/parameters/paperOpt'
        - $ref: '#/components/parameters/orientationOpt'
        - $ref: '#/components/parameters/fontSizeOpt'
        - $ref: '#/components/parameters/renderTableOpt'
        - $ref: '#/components/parameters/renderDetailsOpt'
      responses:
        '200':
          $ref: '#/components/responses/ExportedContent'
        '400':
          $ref: '#/components/responses/Invalid'
components:
  parameters:
    contentTypeOpt:
      name: contentType
      in: query
      description: Overrides the HTTP request's Accept header, since browsers ignore the type attribute in, for example,
        <a href="..." download="filename.pdf" type="application/pdf">filename.pdf</a>. In other words, there's
        no way for an HTML link to force a browser to request the necessary content type. The
        "application/x-research-info-systems" content type is only supported for recordKind="publications".
      required: false
      schema:
        enum:
          - application/pdf
          - text/csv
          - text/html
          - application/x-research-info-systems
        default: application/pdf
    recordKind:
      name: recordKind
      description: The list's record type.
      in: path
      schema:
        type: string
        enum:
          - claims
          - comments
          - declarations
          - groups
          - journals
          - log
          - persons
          - publications
          - publishers
          - quotations
          - topics
          - users
      required: true
    recordIdOpt:
      name: recordId
      description: The identifier of the single record to return. Applicable to all recordKind except logs.
      in: query
      schema:
        type: integer
        format: int64
      required: false
    textOpt:
      name: text
      description: Free text search string. Applicable to all recordKind except logs.
      in: query
      schema:
        type: string
      required: false
    advancedSearchOpt:
      name: advancedSearch
      description: Whether to search text in advanced (boolean) mode, used in conjunction with the text parameter. Applicable to all recordKind except logs.
      in: query
      schema:
        type: boolean
      required: false
    statusOpt:
      name: status
      description: Return only records with this status code. Applicable to all recordKind except logs.
      in: query
      schema:
        type: string
        enum:
          - DRA
          - DEL
          - PUB
          - SUS
      required: false
      style: form
      explode: true
    fromEntityIdOpt:
      name: fromEntityId
      description: The ID of the 'linked-from' master record. Applicable to recordKind claims, declarations, persons, publications, quotations.
      in: query
      schema:
        $ref: '#/components/schemas/id'
      required: false
    fromEntityKindOpt:
      name: fromEntityKind
      description: The kind of the 'linked-from' master record. Applicable to recordKind claims, declarations, persons, publications, quotations.
      in: query
      schema:
        type: string
        enum:
          - CLA
          - DEC
          - PER
          - PUB
          - QUO
          - TOP
      required: false
    toEntityIdOpt:
      name: toEntityId
      description: The ID of the 'linked-to' master record. Applicable to recordKind claims, declarations, persons, publications, quotations.
      in: query
      schema:
        $ref: '#/components/schemas/id'
      required: false
    toEntityKindOpt:
      name: toEntityKind
      description: The kind of the 'linked-to' master record. Applicable to recordKind claims, declarations, persons, publications, quotations.
      in: query
      schema:
        type: string
        enum:
          - CLA
          - DEC
          - PER
          - PUB
          - QUO
          - TOP
      required: false
    topicIdOpt:
      name: topicId
      description: The master topic identifier. Applicable to recordKind claims, declarations, persons, publications, quotations.
      in: query
      schema:
        $ref: '#/components/schemas/id'
      required: false
    recursiveOpt:
      name: recursive
      description: Whether queries including a topicId should recursively include sub-topics. Applicable to recordKind claims, declarations, persons, publications, quotations.
      in: query
      schema:
        type: boolean
      required: false
    targetKindOpt:
      name: targetKind
      description: Restrict to specific target entity kind. Applicable to recordKind comments.
      in: query
      schema:
        type: string
        enum:
          - CLA
          - DEC
          - GRP
          - JOU
          - PER
          - PUB
          - PBR
          - QUO
          - TOP
          - USR
      required: false
    targetIdOpt:
      name: targetId
      description: Restrict to specific target entity ID. Applicable to recordKind comments.
      in: query
      schema:
        $ref: '#/components/schemas/id'
      required: false
    parentIdOpt:
      name: parentId
      description: Restrict to sub-topics thereof or replies to a specific comment. Applicable to recordKind comments, topics.
      in: query
      schema:
        $ref: '#/components/schemas/id'
      required: false
    userIdOpt:
      name: userId
      description: Restrict to logs for a specific user or comments created by a specific user. Applicable to recordKind comments, logs.
      in: query
      schema:
        $ref: '#/components/schemas/id'
      required: false
    fromOpt:
      name: from
      description: Timestamp of first record to include. Applicable to recordKind comments, logs.
      in: query
      schema:
        $ref: '#/components/schemas/datetime'
      required: false
    toOpt:
      name: to
      description: Timestamp of last record to include. Applicable to recordKind comments, logs.
      in: query
      schema:
        $ref: '#/components/schemas/datetime'
      required: false
    entityKindOpt:
      name: entityKind
      description: Restrict to logs for this specific entity kind. Applicable to recordKind logs.
      in: query
      schema:
        type: string
        enum:
          - CLA
          - COM
          - DEC
          - GRP
          - JOU
          - PER
          - PUB
          - PBR
          - QUO
          - TOP
          - USR
      required: false
    entityIdOpt:
      name: entityId
      description: Restrict to logs for this specific entity ID. Applicable to recordKind logs.
      in: query
      schema:
        $ref: '#/components/schemas/id'
      required: false
    transactionKindOpt:
      name: transactionKind
      description: Restrict to logs for this specific transaction kind. Applicable to recordKind logs.
      in: query
      schema:
        type: string
        enum:
          - COM
          - CRE
          - DEL
          - LNK
          - UNL
          - UPD
      required: false
    pageNumberOpt:
      name: pageNumber
      description: The (zero-based) page number to request.
      in: query
      schema:
        type: integer
        format: int32
      required: false
    pageSizeOpt:
      name: pageSize
      description: The page size to use.
      in: query
      schema:
        type: integer
        format: int32
      required: false
    sortOpt:
      name: sort
      description: The sort order(s) to apply, formatted as column-name[ ASC|DESC].
      in: query
      schema:
        type: array
        items:
          type: string
          pattern: "^[a-zA-Z]+(?:\\s+(?:[aA][sS][cC]|[dD][eE][sS][cC]))?$"
      required: false
    columnsOpt:
      name: col
      description: The columns to include.
      in: query
      schema:
        type: array
        items:
          type: string
          pattern: "^[a-zA-Z]+$"
      required: false
    paperOpt:
      name: paper
      description: The size of paper to use (applies only to PDF output)
      in: query
      schema:
        type: string
        enum:
          - A3
          - A4
        default: A4
    orientationOpt:
      name: orientation
      description: The page orientation (applies only to PDF output)
      in: query
      schema:
        type: string
        enum:
          - portrait
          - landscape
        default: portrait
    fontSizeOpt:
      name: fontSize
      description: The font size to use  (applies only to HTML and PDF output)
      in: query
      schema:
        type: integer
        format: int32
        minimum: 8
        maximum: 16
        default: 12
    renderTableOpt:
      name: renderTable
      description: Whether to render a table
      in: query
      schema:
        type: boolean
        default: false
      required: false
    renderDetailsOpt:
      name: renderDetails
      description: Whether to render record details
      in: query
      schema:
        type: boolean
        default: false
      required: false
  schemas:
    id:
      description: A unique object identifier.
      type: integer
      format: int64
    # Unused, because if used, codegen creates a model enum that duplicate existing enums and have unresolvable imports.
    entityKind:
      description: Identifies the kind of tracked entity.
      type: string
      enum:
        - CLA
        - COM
        - DEC
        - GRP
        - JOU
        - PER
        - PUB
        - PBR
        - QUO
        - TOP
        - USR
    # Unused, because if used, codegen creates a model enum that duplicate existing enums and have unresolvable imports.
    linkableEntityKind:
      description: Identifies the kind of linkable entity.
      type: string
      enum:
        - CLA
        - DEC
        - PER
        - PUB
        - QUO
        - TOP
    date:
      description: An ISO 8601 date value.
      type: string
      pattern: "^\\d{4}-\\d{2}-\\d{2}$"
    datetime:
      description: An ISO 8601 date-time value.
      type: string
      pattern: "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}(?:\\.\\d{3})Z?$"
  responses:
    ExportedContent:
      description: The exported list content
      content:
        application/pdf:
          schema:
            type: string
            format: binary
        application/x-research-info-systems:
          schema:
            type: string
        text/csv:
          schema:
            type: string
        text/html:
          schema:
            type: string
    Invalid:
      description: Bad request (e.g., invalid combination of query parameters)
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT